requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
openpyxl==3.1.5
pydantic==2.11.7
asyncpg==0.30.0
python-dotenv==1.0.1